- introduced loss function to decouple loss and optimizer in solver - model saving\restore

+++ src/rbf_network/consts.py
@@ -1,3 +1,3 @@
 import tensorflow as tf
 
-type = tf.float32
+type = tf.float64
